pcap: fix timeout when TPACKET_V3 is used (#15)
Starting with libpcap 1.8.0, the TPACKET_V3 interface will be used
under Linux. This has the side-effect of packet reads not timing out
until a packet is received.

This patch detects when TPACKET_V3 is active and uses poll() on the
capture file descriptor to ensure packet reads complete within
the configured timeout.